refactor(computer): `SandboxInstance` becomes the actor's runtime state
The struct outlived its name. It is no longer an instance in a map that
anyone can lock: it is one computer's runtime state, reachable from its actor
and the sub-tasks that actor spawns and from nowhere else. `ComputerRuntime`,
in `lifecycle/runtime.rs`, beside the actor that owns it.

Mechanical apart from the move; the `Mutex` that remains is what lets a boot
hand a `CowHandle` over mid-flight, which is documented at the type.
